Private method has been created on purpose in order to illustrate inappropriate testing of private methods.
This method is not very complex and can be tested through the public constructor (as it has been done here). It is used as a simplification of a more complex method (with several if statements and/or external libraries for instance).
Moreover, the fact that the class does not apply the Single Responsibility Principle is the reason why the "complex" private method seems to need tests in the first place.